+++ ol/cli/src/node/autopay_view.rs
@@ -1,0 +1,98 @@
+//! `chain_info`
+use anyhow::{bail, Error};
+use chrono::Utc;
+use diem_json_rpc_client::views::OracleUpgradeStateView;
+use diem_types::{
+    account_address::AccountAddress, account_state::AccountState,
+    ol_validators_stats::ValidatorsStatsResource, waypoint::Waypoint,
+};
+use ol_types::{autopay::AutoPayView, validator_config::ValidatorConfigView};
+
+use super::{node::Node, chain_view::ValidatorView};
+use serde::{Deserialize, Serialize};
+use std::{collections::HashMap, convert::TryFrom};
+
+
+#[derive(Clone, Debug, Deserialize, Serialize)]
+///
+pub struct PayeeStats {
+    ///
+    pub address: AccountAddress,
+    ///
+    pub note: String,
+    ///
+    pub balance: f64,
+    ///
+    pub payers: u64,
+    ///
+    pub average_percent: f64,
+    ///
+    pub sum_percentage: u64,
+    ///
+    pub all_percentage: f64,
+}
+
+
+impl Node {
+    /// Get all percentage recurring payees stats
+    pub fn get_autopay_watch_list(&mut self, vals: Vec<ValidatorView>) -> Option<Vec<PayeeStats>> {
+        let mut payees: HashMap<AccountAddress, PayeeSums> = HashMap::new();
+        let mut total: u64 = 0;
+
+        struct PayeeSums {
+            pub amount: u64,
+            pub payers: u64,
+        }
+
+        // iterate over all validators
+        for val in vals.iter() {
+            if let Some(ap) = &val.autopay {
+                // iterate over all autopay instructions
+                let mut val_payees: HashMap<AccountAddress, u64> = HashMap::new();
+                for payment in ap.payments.iter() {
+                    if payment.is_percent_of_change() {
+                        total += payment.amt;
+                        *val_payees.entry(payment.payee).or_insert(0) += payment.amt;
+                    }
+                }
+                // sum payers and amount
+                for (payee, amount) in val_payees.iter() {
+                    let payee_sums = payees.get_mut(&payee);
+                    match payee_sums {
+                        Some(p) => {
+                            p.amount = p.amount + amount;
+                            p.payers = p.payers + 1;
+                        }
+                        None => {
+                            payees.insert(
+                                *payee,
+                                PayeeSums {
+                                    amount: *amount,
+                                    payers: 1,
+                                },
+                            );
+                        }
+                    }
+                }
+            }
+        }
+
+        // collect payees stats
+        let dict = self.load_account_dictionary();
+        let ret = payees
+            .iter()
+            .map(|(payee, stat)| PayeeStats {
+                note: dict.get_note_for_address(*payee),
+                address: *payee,
+                payers: stat.payers,
+                average_percent: stat.amount as f64 / stat.payers as f64,
+                balance: self.get_account_balance(*payee).unwrap_or(0.0),
+                sum_percentage: stat.amount,
+                all_percentage: (stat.amount * 10000) as f64 / total as f64,
+            })
+            .collect();
+        Some(ret)
+    }
+}
+
+
