+++ ol/cli/src/node/chain_view.rs
@@ -4,11 +4,12 @@ use chrono::Utc;
 use diem_json_rpc_client::views::OracleUpgradeStateView;
 use diem_types::{
     account_address::AccountAddress, account_state::AccountState,
-    ol_validators_stats::ValidatorsStatsResource, waypoint::Waypoint,
+    ol_validators_stats::ValidatorsStatsResource, validator_info::ValidatorInfo,
+    waypoint::Waypoint,
 };
 use ol_types::{autopay::AutoPayView, validator_config::ValidatorConfigView};
 
-use super::node::Node;
+use super::{node::Node, dictionary::AccountDictionary, autopay_view::PayeeStats};
 use serde::{Deserialize, Serialize};
 use std::{collections::HashMap, convert::TryFrom};
 
@@ -121,16 +122,14 @@ impl Node {
             let account_state = AccountState::try_from(&account_blob)?;
             let meta = self.client.get_metadata()?;
 
-            let conf_resource = match account_state
-            .get_configuration_resource()? {
+            let conf_resource = match account_state.get_configuration_resource()? {
                 Some(cr) => cr,
                 None => bail!("cannot get configuration resource from chain"),
             };
 
             cs.epoch = conf_resource.epoch();
 
-            let validator_set = match account_state
-            .get_validator_set()? {
+            let validator_set = match account_state.get_validator_set()? {
                 Some(vs) => vs,
                 None => bail!("cannot get validator set resource from chain"),
             };
@@ -138,7 +137,7 @@ impl Node {
             cs.validator_count = validator_set.payload().len() as u64;
 
             // Get vals stats
-            let validators_stats = match account_state.get_validators_stats()?{
+            let validators_stats = match account_state.get_validators_stats()? {
                 Some(vsr) => vsr,
                 None => bail!("could not get validators stats"),
             };
@@ -167,86 +166,13 @@ impl Node {
 
             cs.height = meta.version;
 
-            cs.upgrade = self
-                .client
-                .get_oracle_upgrade_state()?;
+            cs.upgrade = self.client.get_oracle_upgrade_state()?;
 
             let dict = self.load_account_dictionary();
             let validators: Vec<ValidatorView> = validator_set
                 .payload()
                 .iter()
-                .filter_map(|v| {
-                    let full_node_ip = match v.config().fullnode_network_addresses() {
-                        Ok(ips) => {
-                            if ips.len() > 0 {
-                                ips.last().unwrap().to_string()
-                            } else {
-                                "--".to_string()
-                            }
-                        }
-                        Err(_) => "--".to_string(),
-                    };
-                    let validator_ip = match v.config().validator_network_addresses() {
-                        Ok(ips) => {
-                            if ips.len() > 0 {
-                                match ips.get(0) {
-                                    Some(i) => i.seq_num().to_string(),
-                                    None => "--".to_string(),
-                                }
-                                
-                            } else {
-                                "--".to_string()
-                            }
-                        }
-                        Err(_) => "--".to_string(),
-                    };
-                    let ms = match self
-                        .client
-                        .get_miner_state(&v.account_address().clone()) {
-                            Ok(Some(ts)) => ts,
-                            _=> return None,
-                        };
-  
-
-                    let validator_stats = match validators_stats
-                      .get_validator_current_stats(v.account_address().clone()){
-                            Ok(v) => v,
-                            _=> return None,
-                        };
-  
-                    
-                    let val_config = match self.get_validator_config(v.account_address().clone()) {
-                        Ok(v) => v,
-                        _ => return None,
-                    };
-
-                    let autopay = match self.get_autopay_view(v.account_address().clone()){
-                        Ok(v) => v,
-                        _ => return None,
-                    };
-
-                    Some(ValidatorView {
-                        account_address: v.account_address().to_string(),
-                        voting_power: v.consensus_voting_power(),
-                        full_node_ip,
-                        pub_key: v.consensus_public_key().to_string(),
-                        validator_ip,
-
-                        tower_height: ms.verified_tower_height,
-                        tower_epoch: ms.latest_epoch_mining,
-
-                        count_proofs_in_epoch: ms.count_proofs_in_epoch,
-                        epochs_validating_and_mining: ms.epochs_validating_and_mining,
-                        contiguous_epochs_validating_and_mining: ms
-                            .contiguous_epochs_validating_and_mining,
-                        epochs_since_last_account_creation: ms.epochs_since_last_account_creation,
-                        vote_count_in_epoch: validator_stats.vote_count,
-                        prop_count_in_epoch: validator_stats.prop_count,
-                        validator_config: Some(val_config),
-                        autopay: Some(autopay),
-                        note: dict.get_note_for_address(*v.account_address()),
-                    })
-                })
+                .filter_map(|v| self.format_validator_info(v, &dict, &validators_stats).ok())
                 .collect();
 
             cs.validator_view = Some(validators.clone());
@@ -262,65 +188,60 @@ impl Node {
         bail!("could not get chain info")
     }
 
-    /// Get all percentage recurring payees stats
-    pub fn get_autopay_watch_list(&mut self, vals: Vec<ValidatorView>) -> Option<Vec<PayeeStats>> {
-        let mut payees: HashMap<AccountAddress, PayeeSums> = HashMap::new();
-        let mut total: u64 = 0;
-
-        struct PayeeSums {
-            pub amount: u64,
-            pub payers: u64,
-        }
-
-        // iterate over all validators
-        for val in vals.iter() {
-            if let Some(ap) = &val.autopay {
-                // iterate over all autopay instructions
-                let mut val_payees: HashMap<AccountAddress, u64> = HashMap::new();
-                for payment in ap.payments.iter() {
-                    if payment.is_percent_of_change() {
-                        total += payment.amt;
-                        *val_payees.entry(payment.payee).or_insert(0) += payment.amt;
-                    }
-                }
-                // sum payers and amount
-                for (payee, amount) in val_payees.iter() {
-                    let payee_sums = payees.get_mut(&payee);
-                    match payee_sums {
-                        Some(p) => {
-                            p.amount = p.amount + amount;
-                            p.payers = p.payers + 1;
-                        }
-                        None => {
-                            payees.insert(
-                                *payee,
-                                PayeeSums {
-                                    amount: *amount,
-                                    payers: 1,
-                                },
-                            );
-                        }
-                    }
-                }
+    fn format_validator_info(&self, v: &ValidatorInfo, dict: &AccountDictionary, stats: &ValidatorsStatsResource) -> Result<ValidatorView, Error> {
+    let full_node_ip = match v.config().fullnode_network_addresses() {
+        Ok(ips) => {
+            if ips.len() > 0 {
+                ips.last().unwrap().to_string()
+            } else {
+                "--".to_string()
             }
         }
+        Err(_) => "--".to_string(),
+    };
+    let validator_ip = match v.config().validator_network_addresses() {
+        Ok(ips) => {
+            if ips.len() > 0 {
+                match ips.get(0) {
+                    Some(i) => i.seq_num().to_string(),
+                    None => "--".to_string(),
+                }
+            } else {
+                "--".to_string()
+            }
+        }
+        Err(_) => "--".to_string(),
+    };
+    let ms  = self.client.get_miner_state(&v.account_address().clone())?.unwrap();
 
-        // collect payees stats
-        let dict = self.load_account_dictionary();
-        let ret = payees
-            .iter()
-            .map(|(payee, stat)| PayeeStats {
-                note: dict.get_note_for_address(*payee),
-                address: *payee,
-                payers: stat.payers,
-                average_percent: stat.amount as f64 / stat.payers as f64,
-                balance: self.get_account_balance(*payee).unwrap_or(0.0),
-                sum_percentage: stat.amount,
-                all_percentage: (stat.amount * 10000) as f64 / total as f64,
-            })
-            .collect();
-        Some(ret)
-    }
+    let one_val_stat = stats.get_validator_current_stats(v.account_address().clone())?;
+
+    let val_config = self.get_validator_config(v.account_address().clone())?;
+
+    let autopay = self.get_autopay_view(v.account_address().clone())?;
+
+    Ok(ValidatorView {
+        account_address: v.account_address().to_string(),
+        voting_power: v.consensus_voting_power(),
+        full_node_ip,
+        pub_key: v.consensus_public_key().to_string(),
+        validator_ip,
+
+        tower_height: ms.verified_tower_height,
+        tower_epoch: ms.latest_epoch_mining,
+        count_proofs_in_epoch: ms.count_proofs_in_epoch,
+        epochs_validating_and_mining: ms.epochs_validating_and_mining,
+        contiguous_epochs_validating_and_mining: ms.contiguous_epochs_validating_and_mining,
+        epochs_since_last_account_creation: ms.epochs_since_last_account_creation,
+        
+        vote_count_in_epoch: one_val_stat.vote_count,
+        prop_count_in_epoch: one_val_stat.prop_count,
+        validator_config: Some(val_config),
+        autopay: Some(autopay),
+        note: dict.get_note_for_address(*v.account_address()),
+    })
+  }
+
 }
 
 fn calc_config_stats(vals: Vec<ValidatorView>) -> Result<ValsConfigStats, Error> {
@@ -330,19 +251,18 @@ fn calc_config_stats(vals: Vec<ValidatorView>) -> Result<ValsConfigStats, Error>
 
     for val in vals.iter() {
         if let Some(config) = val.validator_config.clone() {
-          if let Some(a) = &val.autopay{
-            if a.payments
-            .iter()
-            .any(|each| each.is_percent_of_change()) {
-              count_autopay += 1;
+            if let Some(a) = &val.autopay {
+                if a.payments.iter().any(|each| each.is_percent_of_change()) {
+                    count_autopay += 1;
+                }
             }
-          }
-          if config.operator_account.is_some() {
-              count_operators += 1;
-          }
-          if config.operator_has_balance.is_some() && config.operator_has_balance.unwrap_or(false) {
-              count_positive_balance += 1;
-          }
+            if config.operator_account.is_some() {
+                count_operators += 1;
+            }
+            if config.operator_has_balance.is_some() && config.operator_has_balance.unwrap_or(false)
+            {
+                count_positive_balance += 1;
+            }
         }
     }
     Ok(ValsConfigStats {
@@ -356,21 +276,3 @@ fn calc_config_stats(vals: Vec<ValidatorView>) -> Result<ValsConfigStats, Error>
     })
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
-///
-pub struct PayeeStats {
-    ///
-    pub address: AccountAddress,
-    ///
-    pub note: String,
-    ///
-    pub balance: f64,
-    ///
-    pub payers: u64,
-    ///
-    pub average_percent: f64,
-    ///
-    pub sum_percentage: u64,
-    ///
-    pub all_percentage: f64,
-}
