+++ ol/cli/src/node/mod.rs
@@ -8,4 +8,5 @@ pub mod client;
 pub mod states;
 pub mod dictionary;
 pub mod refresh_peers;
+pub mod autopay_view;
 // mod transitions;
